+++ lib/maze/option/validator.rb
@@ -2,7 +2,7 @@
 
 require 'yaml'
 require_relative '../option'
-require_relative '../devices'
+require_relative '../browser_stack_devices'
 
 module Maze
   module Option
@@ -15,7 +15,9 @@ module Maze
 
         # Common options
         farm = options[Option::FARM]
-        errors << "--#{Option::FARM} must be either 'bs' or 'local' if provided" if farm && !%w[bs local].include?(farm)
+        if farm && !%w[bs sl local].include?(farm)
+          errors << "--#{Option::FARM} must be 'bs', 'sl' or 'local' if provided"
+        end
 
         begin
           JSON.parse(options[Option::CAPABILITIES])
@@ -25,6 +27,7 @@ module Maze
 
         # Farm specific options
         validate_bs options, errors if farm == 'bs'
+        # TODO validate_sl options, errors if farm == 'sl'
         validate_local options, errors if farm == 'local'
 
         errors
@@ -37,27 +40,28 @@ module Maze
         errors << "BrowserStack local binary '#{bs_local}' not found" unless File.exist? bs_local
 
         # Device
-        bs_browser = options[Option::BS_BROWSER]
-        bs_device = options[Option::BS_DEVICE]
-        if bs_browser.nil? && bs_device.nil?
-          errors << "Either --#{Option::BS_BROWSER} or --#{Option::BS_DEVICE} must be specified"
-        elsif bs_browser
+        test_browser = options[Option::BROWSER]
+        test_device = options[Option::DEVICE]
+        if test_browser.nil? && test_device.nil?
+          errors << "Either --#{Option::BROWSER} or --#{Option::DEVICE} must be specified"
+        elsif test_browser
 
           browsers = YAML.safe_load(File.read("#{__dir__}/../browsers.yml"))
 
-          unless browsers.include? bs_browser
+          unless browsers.include? test_browser
             browser_list = browsers.keys.join ', '
-            errors << "Browser type '#{bs_browser}' unknown on BrowserStack.  Must be one of: #{browser_list}."
+            errors << "Browser type '#{test_browser}' unknown on BrowserStack.  Must be one of: #{browser_list}."
           end
-        elsif bs_device
-          unless Maze::Devices::DEVICE_HASH.key? bs_device
-            errors << "Device type '#{bs_device}' unknown on BrowserStack.  Must be one of #{Maze::Devices::DEVICE_HASH.keys}"
+        elsif test_device
+          unless Maze::BrowserStackDevices::DEVICE_HASH.key? test_device
+            errors << "Device type '#{test_device}' unknown on BrowserStack.  Must be one of #{Maze::BrowserStackDevices::DEVICE_HASH.keys}"
           end
           # App
           app = options[Option::APP]
           if app.nil?
             errors << "--#{Option::APP} must be provided when running on a device"
           else
+            # TODO: What about Sauce Labs URLs?
             unless app.start_with?('bs://')
               app = Maze::Helper.expand_path app
               errors << "app file '#{app}' not found" unless File.exist?(app)